OOPS-56: integrate: done update/create card and update setting, subject

+++ api/src/main/java/com/app/oopsly/api/entity/SubjectEntity.java
@@ -35,7 +35,7 @@ public class SubjectEntity extends Audit {
 
     @Builder.Default private Integer dailyLimit = 20;
     @Builder.Default private Integer newCardsPerDay = 5;
-    @Builder.Default private Integer interval = 1;
+    @Builder.Default private Double interval = 1.0;
 
     @JsonIgnore
     @ManyToOne(fetch = FetchType.LAZY)
